SFR-2289: Reorganizing process file/folder structure
