+++ processes/__init__.py
@@ -1,21 +1,21 @@
-from .covers import CoverProcess
+from .file.covers import CoverProcess
 from .ingest.doab import DOABProcess
 from .ingest.gutenberg import GutenbergProcess
 from .ingest.hathi_trust import HathiTrustProcess
 from .ingest.nypl import NYPLProcess
-from .catalog import CatalogProcess
-from .classify import ClassifyProcess
+from .frbr.catalog import CatalogProcess
+from .frbr.classify import ClassifyProcess
 from .cluster import ClusterProcess
-from .local_development_setup import LocalDevelopmentSetupProcess
-from .seed_local_data import SeedLocalDataProcess
-from .s3_files import S3Process
+from .local_development.local_development_setup import LocalDevelopmentSetupProcess
+from .local_development.seed_local_data import SeedLocalDataProcess
+from .file.s3_files import S3Process
 from .api import APIProcess
 from .ingest.muse import MUSEProcess
 from .ingest.met import METProcess
-from .db_maintenance import DatabaseMaintenanceProcess
-from .db_migration import MigrationProcess
+from .util.db_maintenance import DatabaseMaintenanceProcess
+from .util.db_migration import MigrationProcess
 from .ingest.chicago_isac import ChicagoISACProcess
 from .ingest.u_of_sc import UofSCProcess
 from .ingest.loc import LOCProcess
 from .ingest.u_of_m import UofMProcess
-from .fulfill_url_manifest import FulfillURLManifestProcess
+from .file.fulfill_url_manifest import FulfillURLManifestProcess
